+++ saleor/graphql/product/mutations/product_variant/product_variant_create.py
@@ -5,8 +5,6 @@ import graphene
 from django.core.exceptions import ValidationError
 from django.utils.text import slugify
 
-from saleor.graphql.shop.utils import get_track_inventory_by_default
-
 from .....attribute import AttributeInputType
 from .....attribute import models as attribute_models
 from .....core.tracing import traced_atomic_transaction
@@ -28,6 +26,7 @@ from ....core.types import BaseInputObjectType, NonNullList, ProductError
 from ....core.utils import get_duplicated_values
 from ....meta.inputs import MetadataInput
 from ....plugins.dataloaders import get_plugin_manager_promise
+from ....shop.utils import get_track_inventory_by_default
 from ....warehouse.types import Warehouse
 from ...types import ProductVariant
 from ...utils import (
@@ -62,7 +61,7 @@ class ProductVariantInput(BaseInputObjectType):
         description=(
             "Determines if the inventory of this variant should be tracked. If false, "
             "the quantity won't change when customers buy this item."
-            "If the field is not provided, Shop.trackInventoryByDefault will be used."
+            "If the field is not provided, `Shop.trackInventoryByDefault` will be used."
         )
     )
     weight = WeightScalar(description="Weight of the Product Variant.", required=False)
@@ -298,14 +297,17 @@ class ProductVariantCreate(ModelMutation):
             )
 
     @classmethod
-    def save(cls, info: ResolveInfo, instance, cleaned_input):
-        new_variant = instance.pk is None
+    def set_track_inventory(cls, _info, instance, cleaned_input):
         instance.track_inventory = (
             get_track_inventory_by_default()
             if cleaned_input.get("track_inventory") is None
             else cleaned_input.get("track_inventory")
         )
-        instance.save()
+
+    @classmethod
+    def save(cls, info: ResolveInfo, instance, cleaned_input):
+        new_variant = instance.pk is None
+        cls.set_track_inventory(info, instance, cleaned_input)
         with traced_atomic_transaction():
             instance.save()
             if not instance.product.default_variant: