+++ saleor/graphql/shop/mutations/shop_settings_update.py
@@ -31,7 +31,7 @@ class ShopSettingsInput(graphene.InputObjectType):
     track_inventory_by_default = graphene.Boolean(
         description=(
             "This field is used as a default value for "
-            "ProductVariant.trackInventory."
+            "`ProductVariant.trackInventory`."
         )
     )
     default_weight_unit = WeightUnitsEnum(description="Default weight unit.")